+++ app/javascript/gobierto_visualizations/modules/subsidies_controller.js
@@ -9,7 +9,7 @@ import {
 import Vue from "vue";
 import VueRouter from "vue-router";
 import { getRemoteData, sortByField, calculateSumMeanMedian } from "../webapp/lib/utils";
-import { EventBus } from "../webapp/mixins/event_bus";
+import { EventBus } from "../webapp/lib/mixins/event_bus";
 // import { checkAndReportAccessibility } from "lib/vue/accesibility";
 
 // if (Vue.config.devtools) {
