+++ app/javascript/gobierto_visualizations/webapp/components/SearchFilter.vue
@@ -21,7 +21,7 @@
 </template>
 
 <script>
-import { EventBus } from "../mixins/event_bus";
+import { EventBus } from "../lib/mixins/event_bus";
 export default {
   name: 'SearchFilter',
   props: {
