+++ app/javascript/gobierto_visualizations/webapp/components/TreeMap.vue
@@ -32,7 +32,7 @@ import { sumDataByGroupKey, slugString } from "../lib/utils";
 import { easeLinear } from 'd3-ease'
 import { mean, median } from "d3-array";
 import { money } from "lib/vue/filters";
-import { EventBus } from "../mixins/event_bus";
+import { EventBus } from "../lib/mixins/event_bus";
 
 const d3 = { select, selectAll, treemap, stratify, mouse, easeLinear, mean, median }
